:fire: :bug: Move the images out of the database.
Putting them in the database in the first place was a big mistake.
Not only is there no point in doing it, but the performances that you
would theoretically get out of it would be horrible, the support for
blob types in MySQL leaves to be desired and PHP's PDOs greatly lack
support for that (see https://bugs.php.net/bug.php?id=40913).

After numerous failed attempt on my side and on @kalharko's, I suggest
we simply move away from blobs and use fs-stored images.